+++ track.py
@@ -1,6 +1,6 @@
 #!/bin/python3
 
-from scapy.all import AsyncSniffer, Ether
+from scapy.all import AsyncSniffer, Ether, UDP
 from threading import Lock, Event, Timer
 import concurrent.futures
 import requests
@@ -24,6 +24,7 @@ class Tracker(object):
         self.token = token
         self.entities = self.config['entities']
         logging.info(f"got entities {self.entities}")
+        self.absence_timeout = self.config['absence_timeout']
 
         logging.info("setting up websocket")
         self.ws = WebSocketHa(config['url']) if config.get('url') else WebSocketHa('ws://supervisor/core/websocket')
@@ -31,11 +32,14 @@ class Tracker(object):
         self.ws.auth(self.token)
 
         self.cache_timeout = self.config['timeout']
-        self.filter = 'udp dst port 67 and udp[248:1] = 0x35 and udp[249:1] = 0x1 and udp[250:1] = 0x3' # DHCP Request
-        self.cache = dict()
+        self.filter = '(udp dst port 67 and udp[248:1] = 0x35 and udp[249:1] = 0x1 and udp[250:1] = 0x3) ' # DHCP Request
+        self.filter += ' '.join([f'or (arp and ether src {entity["mac"]})' for entity in self.entities])
+        now = time.time()
+        self.cache = dict(map(lambda entity: (entity['mac'], now), self.entities))
         self.pkts = list()
         self.__lock = Lock()
         self.__event = Event()
+        logging.debug(f"using filter -> {self.filter}")
         self.sniffer = AsyncSniffer(filter=self.filter, prn=lambda x: self.__insert_packet(x), store=0)
 
     def __insert_packet(self, pkt):
@@ -43,6 +47,15 @@ class Tracker(object):
         self.pkts.append(pkt)
         self.__lock.release()
         self.__event.set()
+
+    def check_absence(self):
+        now = time.time()
+        for entity in self.entities:
+            mac = entity['mac']
+            if mac in self.cache and now - self.cache[mac] > self.absence_timeout:
+                logging.info(f"entity {entity['name']} left home")
+                self.see(entity['entity'], entity['name'], mac, "not_home")
+                self.cache.pop(entity['mac'])
 
     def track(self):
         logging.info("tracker started")
@@ -52,10 +65,12 @@ class Tracker(object):
         while True:
 
             # keep connection alive every 5 seconds
+            # check absence of tracked devices every 5 seconds
             cur_time = time.time()
             delta = cur_time - last_keepalive
             if delta > interval: 
                 self.ws.ping()
+                self.check_absence()
                 last_keepalive = cur_time
                 delta = 0
 
@@ -68,8 +83,17 @@ class Tracker(object):
                 for pkt in pkts:
                     self.handle_packet(pkt)
 
-            Timer(interval - delta, lambda: self.__event.set()).start()
+            # start a timer for ping
+            timer = Timer(interval - delta, lambda: self.__event.set())
+            timer.start()
+
+            # wait for either timer or a packet to come
             self.__event.wait()
+
+            # make sure timer is stopped
+            timer.cancel()
+
+            # clear the event
             self.__event.clear()
 
         self.sniffer.stop()
@@ -80,8 +104,9 @@ class Tracker(object):
     def handle_packet(self, pkt):
         logging.info(pkt.summary())
         mac = pkt[Ether].src
-        if self.cache_invalid(mac):
-            self.notify(pkt[Ether].src)
+        if self.should_track_mac(mac):
+            if pkt.getlayer(UDP) and mac not in self.cache:
+                self.notify(mac)
             self.cache[mac] = time.time()
 
     def should_track_mac(self, mac):
@@ -94,14 +119,10 @@ class Tracker(object):
 
     def notify(self, mac):
         logging.info(f"notifying {mac}")
-        if self.should_track_mac(mac):
-            entity = self.get_entity_by_mac(mac)
-            dev_id = entity['entity'].split('.')[1]
-            name = entity['name']
-
-            # since we do not have absence detection, do this to trigger state change
-            self.see(dev_id, name, mac, "not_home")
-            self.see(dev_id, name, mac, "home")
+        entity = self.get_entity_by_mac(mac)
+        dev_id = entity['entity'].split('.')[1]
+        name = entity['name']
+        self.see(dev_id, name, mac, "home")
 
     def see(self, dev_id, name, mac, location):
         try: